requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.4
distro==1.9.0
exceptiongroup==1.3.1
filetype==1.2.0
google-auth==2.48.0
google-genai==1.61.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.8
langchain-core==1.2.8
langchain-google-genai==4.2.0
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.8
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.0
websockets==15.0.1
xxhash==3.6.0
zstandard==0.25.0